Add keymap.el to generate asset from Agda emacs mode
The script loads agda-mode.el and generates both a JSON and JavaScript
keymap/query file. This removes the need for a hand-edited/extracted
copy of Emacs agda-mode source, which makes it easier to keep up-to-date
with the original mappings. Additionally this includes the inherited
mappings from `latin-ltx` without need for copying them.

The script can be run with `npm run keymap-gen`, or run with custom
arguments like so:
```
./keymap.el generate --agda-emacs-mode-dir ~/Code/agda/src/data/emacs-mode
```